CRUDS Addon

+++ database/migrations/2023_09_15_052136_create_addons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('mst_add_on', function (Blueprint $table) {
             $table->id();
             $table->string('add_on_name');
-            $table->float('price');
+            $table->double('price',20,0);
             $table->boolean('is_active');
             $table->string('created_by');
             $table->string('updated_by')->nullable();
